+++ client/src/constants.ts
@@ -1,1 +1,2 @@
 export const wollokLSPExtensionCode = 'wollokLSP'
+export const wollokLSPExtensionId = 'wollok-lsp-ide'