+++ client/src/extension.ts
@@ -20,6 +20,7 @@ import {
 } from 'vscode-languageclient/node'
 import { subscribeWollokCommands } from './commands'
 import { allWollokFiles } from './utils'
+import { wollokLSPExtensionId } from './constants'
 
 let client: LanguageClient
 
@@ -48,7 +49,7 @@ export function activate(context: ExtensionContext): void {
     // Register the server for Wollok documents
     documentSelector: [{ scheme: 'file', language: 'wollok' }],
     synchronize: {
-      configurationSection: 'wollok-lsp-ide',
+      configurationSection: wollokLSPExtensionId,
       // Notify the server about file changes to '.clientrc files contained in the workspace
       fileEvents: workspace.createFileSystemWatcher('**/.clientrc'),
     },
@@ -59,7 +60,7 @@ export function activate(context: ExtensionContext): void {
 
   // Create the language client and start the client.
   client = new LanguageClient(
-    'wollok-lsp-ide',
+    wollokLSPExtensionId,
     'Wollok',
     serverOptions,
     clientOptions,
@@ -91,14 +92,12 @@ export function activate(context: ExtensionContext): void {
 }
 
 export function deactivate(): Thenable<void> | undefined {
-  if (!client) {
-    return undefined
-  }
-  return client.stop()
+  return client?.stop()
 }
 
 async function validateWorkspace() {
   const uris = await allWollokFiles()
+  await client.sendRequest(`WORKSPACE_URI:${workspace.workspaceFolders[0].uri}`)
   for (const uri of uris) {
     // Force 'change' on document for server tracking
     const textDoc = await workspace.openTextDocument(uri)