+++ client/src/test/hover.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert'
 import { commands, Hover, MarkdownString, Position, Range, Uri } from 'vscode'
-import { activate, getDocumentURI } from './helper'
+import { activate, getDocumentURI, setConfiguration } from './helper'
 
 /** ATTENTION
  * These tests are NOT ATOMIC, they depend on each other, order matters. (Resolve TODO)
@@ -8,7 +8,8 @@ import { activate, getDocumentURI } from './helper'
 suite('Should display on hover', () => {
   const hoverURI = getDocumentURI('hover.wlk')
 
-  test('hover field', async () => {
+  test('hover field with type info', async () => {
+    await setConfiguration('typeSystem.enabled', true)
     await testHover(
       hoverURI,
       new Position(1, 8),
@@ -17,7 +18,22 @@ suite('Should display on hover', () => {
           new MarkdownString('\n```text\nField: Number\n```\n'),
           new MarkdownString('\n```wollok\nconst x = 2\n```\n'),
         ],
-        new Range(new Position(1, 2), new Position(2, 0))
+        new Range(new Position(1, 2), new Position(1, 13))
+      )
+    )
+  })
+
+  test('hover field without type info', async () => {
+    await setConfiguration('typeSystem.enabled', false)
+    await testHover(
+      hoverURI,
+      new Position(1, 8),
+      new Hover(
+        [
+          new MarkdownString('\n```text\nField\n```\n'),
+          new MarkdownString('\n```wollok\nconst x = 2\n```\n'),
+        ],
+        new Range(new Position(1, 2), new Position(1, 13))
       )
     )
   })