+++ server/src/functionalities/symbols.ts
@@ -1,16 +1,12 @@
 import { DocumentSymbol, DocumentSymbolParams, SymbolKind, WorkspaceSymbol, WorkspaceSymbolParams } from 'vscode-languageserver'
 import { Environment, Field, Method, Module, Node, Package, Program, Test, Variable } from 'wollok-ts'
-import { packageFromURI, toVSCRange } from '../utils/text-documents'
-import { workspacePackage } from '../utils/vm/wollok'
 import { logger } from '../utils/logger'
+import { packageFromURI, toVSCRange, uriFromRelativeFilePath } from '../utils/text-documents'
+import { projectPackages } from '../utils/vm/wollok'
 
 type Symbolyzable = Program | Test | Module | Variable | Field | Method | Test
 
 export const documentSymbols = (environment: Environment) => (params: DocumentSymbolParams): DocumentSymbol[] => {
-  // ToDo this is a temporal fix for https://github.com/uqbar-project/wollok-lsp-ide/issues/61
-  if (!workspacePackage(environment)) {
-    return []
-  }
   const document = packageFromURI(params.textDocument.uri, environment)
   if (!document){
     logger.error('Could not produce symbols: document not found')
@@ -28,9 +24,10 @@ const documentSymbolsFor = (document: Package): DocumentSymbol[] =>
   (document.members.filter(isSymbolyzable) as Symbolyzable[]).map(documentSymbol)
 
 const workspaceSymbolsFor = (environment: Environment, query: string): WorkspaceSymbol[] =>
-  workspacePackage(environment).descendants.filter(isSymbolyzable)
+  projectPackages(environment)
+    .flatMap(_package => _package.descendants).filter(isSymbolyzable)
     .filter(node => node.sourceFileName && node.sourceMap)
-    .filter(node => node.name?.toLowerCase().includes(query.toLowerCase()))
+    .filter(node => node.name?.toLowerCase().includes(query.toLowerCase())) // TODO: can we simplify to a single filter? Maybe extract the function
     .map(workspaceSymbol)
 
 
@@ -49,13 +46,12 @@ const documentSymbol = (node: Symbolyzable): DocumentSymbol => {
 const workspaceSymbol = (node: Symbolyzable): WorkspaceSymbol => WorkspaceSymbol.create(
   node.name!,
   symbolKind(node),
-  node.sourceFileName!,
+  uriFromRelativeFilePath(node.sourceFileName!),
   toVSCRange(node.sourceMap!)
 )
 
 const isSymbolyzable = (node: Node): node is Symbolyzable =>
   [Program, Test, Module, Variable, Field, Method].some(t => node.is(t))
-
 
 const symbolKind = (node: Node): SymbolKind => {
   switch (node.kind) {