+++ server/src/utils/text-documents.ts
@@ -1,6 +1,8 @@
+import fs from 'fs'
+import path from 'path'
 import { Location, Position, Range, TextDocumentIdentifier, TextDocumentPositionParams } from 'vscode-languageserver'
 import { TextDocument } from 'vscode-languageserver-textdocument'
-import { Environment, Node, Package, SourceIndex, SourceMap } from 'wollok-ts'
+import { Environment, FileContent, Node, PROGRAM_FILE_EXTENSION, Package, SourceIndex, SourceMap, TEST_FILE_EXTENSION, WOLLOK_FILE_EXTENSION } from 'wollok-ts'
 
 // TODO: Refactor
 const include = (node: Node, { position, textDocument: { uri } }: TextDocumentPositionParams) => {
@@ -8,7 +10,7 @@ const include = (node: Node, { position, textDocument: { uri } }: TextDocumentPo
   if (node.kind === 'Package') {
     return uri.includes(node.sourceFileName)
   }
-  if(!node.sourceMap) return false
+  if (!node.sourceMap) return false
 
   const startPosition = toVSCPosition(node.sourceMap.start)
   const endPosition = toVSCPosition(node.sourceMap.end)
@@ -23,41 +25,39 @@ export const between = (pointer: Position, start: Position, end: Position): bool
   const { line: lineStart, character: charStart } = start
   const { line: lineEnd, character: charEnd } = end
 
-  if(lineStart === lineEnd && linePointer === lineStart) {
+  if (lineStart === lineEnd && linePointer === lineStart) {
     return charPointer >= charStart && charPointer <= charEnd
   }
 
   return linePointer > lineStart
     && linePointer < lineEnd
     || (linePointer === lineStart
-    && charPointer >= charStart
-    || linePointer === lineEnd
-    && charPointer <= charEnd)
+      && charPointer >= charStart
+      || linePointer === lineEnd
+      && charPointer <= charEnd)
 }
 
 export const getNodesByPosition = (environment: Environment, textDocumentPosition: TextDocumentPositionParams): Node[] => {
   return environment.descendants.filter(node => !!node.sourceFileName && include(node, textDocumentPosition))
 }
 
-export const toVSCPosition = (position: SourceIndex): Position => {
-  const max0 = (n: number) => n < 0 ? 0 : n
-
-  return Position.create(
-    max0(position.line - 1),
-    max0(position.column - 1)
+export const toVSCPosition = (position: SourceIndex): Position =>
+  Position.create(
+    Math.max(0, position.line - 1),
+    Math.max(0, position.column - 1)
   )
-}
+
 
 export const toVSCRange = (sourceMap: SourceMap): Range =>
   Range.create(toVSCPosition(sourceMap.start), toVSCPosition(sourceMap.end))
 
 export const nodeToLocation = (node: Node): Location => {
-  if(!node.sourceMap || !node.sourceFileName){
+  if (!node.sourceMap || !node.sourceFileName) {
     throw new Error('No source map found for node')
   }
 
   return {
-    uri: node.sourceFileName!,
+    uri: uriFromRelativeFilePath(node.sourceFileName!),
     range: toVSCRange(node.sourceMap),
   }
 }
@@ -76,7 +76,8 @@ export function trimIn(range: Range, textDocument: TextDocument): Range {
 }
 
 export const packageFromURI = (uri: string, environment: Environment): Package | undefined => {
-  const sanitizedURI = uri.replace('file:///', '')
+  const sanitizedURI = relativeFilePath(uri)
+  // TODO: Use projectFQN ?
   return environment.descendants.find(node => node.is(Package) && node.fileName === sanitizedURI) as Package | undefined
 }
 
@@ -84,19 +85,20 @@ export const packageToURI = (pkg: Package): string => fileNameToURI(pkg.fileName
 
 export const fileNameToURI = (fileName: string): string => `file:///${fileName}`
 
-export const getWollokFileExtension = (uri: string): 'wlk' | 'wpgm' | 'wtest' => {
+export const getWollokFileExtension = (uri: string): typeof WOLLOK_FILE_EXTENSION | typeof PROGRAM_FILE_EXTENSION | typeof TEST_FILE_EXTENSION => {
   const extension = uri.split('.').pop()
-  if(!extension) throw new Error('Could not determine file extension')
+  if (!extension) throw new Error('Could not determine file extension')
 
   switch(extension) {
-    case 'wlk':
-    case 'wpgm':
-    case 'wtest':
+    case WOLLOK_FILE_EXTENSION:
+    case PROGRAM_FILE_EXTENSION:
+    case TEST_FILE_EXTENSION:
       return extension
     default:
-      throw new Error('Invalid file extension')
+      throw new Error(`Invalid file extension: ${extension}`)
   }
 }
+
 export function cursorNode(
   environment: Environment,
   position: Position,
@@ -106,4 +108,37 @@ export function cursorNode(
     position,
     textDocument,
   }).reverse()[0]
+}
+
+/** URI */
+
+export let WORKSPACE_URI = ''
+
+export const setWorkspaceUri = (uri: string): void => {
+  WORKSPACE_URI = uri
+}
+
+export const relativeFilePath = (absoluteURI: string): string => {
+  return absoluteURI.replaceAll(WORKSPACE_URI + '/', '')
+}
+
+export const uriFromRelativeFilePath = (relativeURI: string): string => {
+  return WORKSPACE_URI + '/' + relativeURI
+}
+
+export const documentToFile = (doc: TextDocument): FileContent => ({
+  name: relativeFilePath(doc.uri),
+  content: doc.getText(),
+})
+
+export const isNodeURI = (node: Node, uri: string): boolean => node.sourceFileName == relativeFilePath(uri)
+
+export const findPackageJSON = (uri: string): string => {
+  let baseUri = uri
+  while (!fs.existsSync(baseUri + path.sep + 'package.json') && baseUri) {
+    const lastIndex = baseUri.lastIndexOf(path.sep)
+    if (!lastIndex) return ''
+    baseUri = baseUri.slice(0, lastIndex)
+  }
+  return baseUri
 }